refactor: test_canister is now test_canister_rs
