+++ scripts/ci_cd_test_integration.sh
@@ -19,8 +19,8 @@ GATEWAY_PRINCIPAL=$(cargo run -q -p scripts --bin principal_from_key_pair "./dat
 echo "Gateway principal: $GATEWAY_PRINCIPAL"
 
 echo "Deploying test canister..."
-cd tests/test_canister
-dfx deploy test_canister --no-wallet --argument "(opt \"$GATEWAY_PRINCIPAL\")"
+cd tests/test_canister_rs
+dfx deploy test_canister_rs --no-wallet --argument "(opt \"$GATEWAY_PRINCIPAL\")"
 # generate test_canister JS declarations
 npm run generate
 # source the canister id from the generated .env file
@@ -29,7 +29,7 @@ source .env
 cd ../integration
 
 echo "Running integration test..."
-WS_GATEWAY_URL=ws://127.0.0.1:8080 IC_URL=http://127.0.0.1:4943 FETCH_IC_ROOT_KEY=true DFX_NETWORK=local TEST_CANISTER_ID=$CANISTER_ID_TEST_CANISTER npm test
+WS_GATEWAY_URL=ws://127.0.0.1:8080 IC_URL=http://127.0.0.1:4943 FETCH_IC_ROOT_KEY=true DFX_NETWORK=local TEST_CANISTER_ID=$CANISTER_ID_TEST_CANISTER_RS npm test
 
 echo "Stopping gateway..."
 kill $gateway_pid
