+++ scripts/install_test_dependencies.sh
@@ -7,7 +7,7 @@ cd integration
 npm install
 
 # install test_canister dependencies
-cd ../test_canister
+cd ../test_canister_rs
 
 rustup target add wasm32-unknown-unknown
