+++ tests/integration/.env.example
@@ -1,6 +1,6 @@
 WS_GATEWAY_URL=ws://127.0.0.1:8080
 IC_URL=http://127.0.0.1:4943
-# if testing against the local replica, read it from ../test_canister/.env
+# if testing against the local replica, read it from ../test_canister_rs/.env
 TEST_CANISTER_ID='bkyz2-fmaaa-aaaaa-qaaaq-cai'
 FETCH_IC_ROOT_KEY=true
 # by setting this to 'ic', the IC_URL must be set to 'https://icp0.io' and FETCH_IC_ROOT_KEY must be set to 'false'
