+++ tests/integration/src/utils/actors.ts
@@ -1,6 +1,6 @@
 import { HttpAgent, HttpAgentOptions } from "@dfinity/agent";
 import environment from "./environment";
-import type { _SERVICE } from "../../../test_canister/src/declarations/test_canister/test_canister.did";
+import type { _SERVICE } from "../../../test_canister_rs/src/declarations/test_canister_rs/test_canister_rs.did";
 
 const commonAgentOptions: HttpAgentOptions = {
   host: environment.IC_URL,
