+++ tests/test_canister_rs/Cargo.toml
@@ -1,0 +1,4 @@
+[workspace]
+members = [
+    "src/test_canister_rs",
+]
