+++ tests/test_canister_rs/canister_ids.json
@@ -1,5 +1,5 @@
 {
-  "test_canister": {
+  "test_canister_rs": {
     "ic": "on7gu-ciaaa-aaaao-a2fxa-cai"
   }
 }
